+++ verl_omni/models/diffusers/__init__.py
@@ -11,15 +11,10 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-import os
 
-with open(os.path.join(os.path.dirname(os.path.abspath(__file__)), "version/version")) as f:
-    __version__ = f.read().strip()
+from .flash_attention_3 import apply_flash_attention_3_varlen_hub_fix
+from .qwen_image import apply_qwen_image_ulysses_mask_fix
 
+apply_flash_attention_3_varlen_hub_fix()  # apply FA3 varlen hub fix universally
 
-# Apply model patches and auto-register pipelines / rollout / reward loop / engines
-import verl_omni.models  # noqa: E402, F401
-import verl_omni.pipelines  # noqa: E402, F401
-import verl_omni.reward_loop  # noqa: E402, F401
-import verl_omni.workers.engine  # noqa: E402, F401
-import verl_omni.workers.rollout  # noqa: E402, F401
+__all__ = ["apply_flash_attention_3_varlen_hub_fix", "apply_qwen_image_ulysses_mask_fix"]
